git-svn-id: svn://tug.org/texlive/trunk/Build/source@4666 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/web2c/xetexdir/XeTeX_mac.c
@@ -162,7 +162,7 @@ DoAtsuiLayout(void* p, int justify)
 		lsDelta -= lsUnit;
 
 	native_glyph_count(node) = realGlyphCount;
-	native_glyph_info_ptr(node) = (long)glyph_info;
+	native_glyph_info_ptr(node) = glyph_info;
 	
 	if (!justify)
 		node_width(node) = FixedPStoTeXPoints(layoutRec[count-1].realPos) + lsDelta;
